bugfix: turn on org-mode in temp buffers
The --update- functions edit the contents of an org-noter "notes" file in a
temporary buffer.  The changes here turn on org-mode if it was not already on.

The `until...` statement guarantees that org-mode is ON after this bit of code.
In the unlikely scenario where the buffer is already in org-mode (prehaps with a
-*- Org -*- on the top line?), the call to (org-mode) would be suppressed, since
in this case it would turn OFF org-mode.